+++ src/Sdk/Endpoints/Endpoint.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Lettermint\Laravel\Sdk\Endpoints;
+
+use Lettermint\Laravel\Sdk\Client\HttpClient;
+
+abstract class Endpoint
+{
+    protected HttpClient $httpClient;
+
+    public function __construct(HttpClient $httpClient)
+    {
+        $this->httpClient = $httpClient;
+    }
+}