+++ src/Sdk/Exceptions/InvalidSignatureException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Lettermint\Laravel\Sdk\Exceptions;
+
+class InvalidSignatureException extends WebhookVerificationException {}