+++ src/Sdk/Exceptions/TimestampToleranceException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Lettermint\Laravel\Sdk\Exceptions;
+
+class TimestampToleranceException extends WebhookVerificationException {}