+++ src/Sdk/Exceptions/WebhookVerificationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Lettermint\Laravel\Sdk\Exceptions;
+
+use Exception;
+
+class WebhookVerificationException extends Exception {}